requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
asyncpg==0.30.0
redis==6.2.0
pydantic==2.11.7
geoalchemy2==0.17.1
geopandas==1.1.1
numpy==2.2.6
pandas==2.3.1
scikit-learn==1.7.1
shapely==2.1.1
pyproj==3.7.1
prophet==1.1.7
reportlab==4.4.3
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
requests==2.32.3
psycopg2-binary==2.9.10
xgboost==3.2.0
openpyxl==3.1.5
pykrige==1.7.2
gstools==1.7.0
rasterio==1.4.3
